feat: support transformers benchmark
