+++ test/bench/backends/__init__.py
@@ -1,0 +1,5 @@
+from .infinilm import InfiniLMBenchmark
+from .transformers import TransformersBenchmark
+from .vllm import VLLMBenchmark
+
+__all__ = ["InfiniLMBenchmark", "TransformersBenchmark", "VLLMBenchmark"]
